Hardwaretest

+++ include/moduleDistance.h
@@ -64,7 +64,7 @@ class moduleDistance {
 		bool Debug();
 		bool Debug(bool debug);
 		uint32 CalcCycle();
-		uint32 CalcCycle(uint8 calcCycle);
+		uint32 CalcCycle(uint32 calcCycle);
 	private:
 		uint16 volumeLast;
 		unsigned long publishVolumeLast;
