+++ include/moduleMoisture.h
@@ -61,7 +61,7 @@ class moduleMoisture {
 		bool Debug();
 		bool Debug(bool debug);
 		uint32 CalcCycle();
-		uint32 CalcCycle(uint8 calcCycle);
+		uint32 CalcCycle(uint32 calcCycle);
 	private:
 		int moistureLast;
 		unsigned long publishMoistureLast;
