+++ include/moduleRpm.h
@@ -55,7 +55,7 @@ class moduleRpm {
 		bool Debug();
 		bool Debug(bool debug);
 		uint32 CalcCycle();
-		uint32 CalcCycle(uint8 calcCycle);
+		uint32 CalcCycle(uint32 calcCycle);
 	private:
 		short rpmLast;
 		unsigned long publishRpmLast;
